# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/contests/{number}": {
      "get": {
        "operationId": "ContestsController_detail",
        "parameters": [
          {
            "name": "number",
            "required": true,
            "in": "path",
            "description": "Contest number",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contest detail"
          },
          "404": {
            "description": "Contest not found"
          }
        },
        "summary": "Get contest detail",
        "tags": [
          "contests"
        ]
      }
    },
    "/contests/{number}/scoreboard": {
      "get": {
        "operationId": "ContestsController_scoreboard",
        "parameters": [
          {
            "name": "number",
            "required": true,
            "in": "path",
            "description": "Contest number",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public scoreboard"
          },
          "404": {
            "description": "Contest not found"
          }
        },
        "summary": "Get contest scoreboard",
        "tags": [
          "contests"
        ]
      }
    },
    "/contests/{number}/scoreboard/full": {
      "get": {
        "operationId": "ContestsController_scoreboardFull",
        "parameters": [
          {
            "name": "number",
            "required": true,
            "in": "path",
            "description": "Contest number",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full scoreboard with all participants"
          },
          "403": {
            "description": "Admin access required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get full contest scoreboard (admin only)",
        "tags": [
          "contests"
        ]
      }
    },
    "/contests/{number}/register": {
      "post": {
        "operationId": "ContestsController_register",
        "parameters": [
          {
            "name": "number",
            "required": true,
            "in": "path",
            "description": "Contest number",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Registered successfully"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Contest not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register for a contest",
        "tags": [
          "contests"
        ]
      }
    },
    "/contests/{number}/verify-password": {
      "post": {
        "operationId": "ContestsController_verifyPassword",
        "parameters": [
          {
            "name": "number",
            "required": true,
            "in": "path",
            "description": "Contest number",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyContestPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password verified"
          },
          "401": {
            "description": "Invalid password"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify contest password for private contests",
        "tags": [
          "contests"
        ]
      }
    },
...
    "/admin/contests/{id}": {
      "get": {
        "operationId": "AdminContestsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Contest ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contest detail"
          },
          "404": {
            "description": "Contest not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get contest detail by ID (admin)",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminContestsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Contest ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contest updated"
          },
          "404": {
            "description": "Contest not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a contest",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminContestsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Contest ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contest deleted"
          },
          "404": {
            "description": "Contest not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a contest",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "EpochJudge API",
    "description": "纪元 — 开源 Online Judge 平台 API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "VerifyContestPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        },
        "required": [
          "password"
        ]
      },
...
      "UpdateContestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "judgeMode": {
            "type": "string",
            "enum": [
              "OI",
              "ACM"
            ]
          },
          "startAt": {
            "type": "string"
          },
          "endAt": {
            "type": "string"
          },
          "freezeAt": {
            "type": "object",
            "nullable": true
          },
          "accessPassword": {
            "type": "object",
            "nullable": true
          },
          "problemIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}